# Cache regexp compilations during string validation.