If there are a lot of routing prefixes to be validated, regexp compilation can become a performance bottleneck.

Aside: This could be because Go's regexp implementation might be relatively slow compared to other languages: https://github.com/golang/go/issues/11646