@wenovus 2021-04-09
790c26b's benchmark test was CPU-profiled using "runtime/pprof", and I see these results for the relevant functions:

The reason why I kept n small for the benchmark tests is that I found other functions (i.e. reflect's MapKeys()) was scaling even worse than the extra compile operations. In order to see a sizeable difference in the benchmark test, I chose a smaller number (too small also doesn't work).

### without optimization
```
BenchmarkPOSIXPattern-12              13          89135241 ns/op
PASS
ok      github.com/openconfig/ygot/ytypes/schema_tests  4.972s
```
#### Running with n=5000 for ipv6-prefix to get a good pprof profile:
```
Duration: 3.64s, Total samples = 5.92s (162.48%)
flat  flat%   sum%        cum   cum%
         0     0% 90.20%      1.08s 18.24%  regexp.CompilePOSIX (inline)
         0     0% 90.20%      1.08s 18.24%  regexp.compile
         0     0% 90.20%      0.62s 10.47%  github.com/openconfig/ygot/ytypes.ValidateStringRestrictions
         0     0% 90.20%      1.12s 18.92%  github.com/openconfig/ygot/ytypes.validateString
         0     0% 90.20%      0.49s  8.28%  github.com/openconfig/ygot/ytypes.validateStringSchema
(*Regexp).MatchString` didn't appear as a profile entry.
```

### with optimization
```
BenchmarkPOSIXPattern-12              19          57036685 ns/op
PASS
ok      github.com/openconfig/ygot/ytypes/schema_tests  4.865s
```
#### Running with n=5000 for ipv6-prefix to get a good pprof profile:
```
Duration: 2.33s, Total samples = 3.44s (147.72%)
         0     0% 92.44%      0.02s  0.58%  runtime.lock (inline)
         0     0% 92.44%      0.02s  0.58%  runtime.lockWithRank (inline)
         0     0% 92.44%      0.05s  1.45%  github.com/openconfig/ygot/ytypes.ValidateStringRestrictions
         0     0% 92.44%      0.05s  1.45%  github.com/openconfig/ygot/ytypes.validateString
`validateStringSchema` didn't appear as a profile entry.
         0     0% 92.44%      0.03s  0.87%  regexp.(*Regexp).MatchString (inline)
```

From the pprof results, it appears the locking operations (in the absence of contention), does not add any significant overhead as compared with the validation operations themselves, which take much less time than the unoptimized version.

So, I think the global cache makes sense as an implementation as-is.