@wenovus 2021-04-09
f96e04f has the same performance:
BenchmarkPOSIXPattern-12              13          89388300 ns/op
PASS
ok      github.com/openconfig/ygot/ytypes/schema_tests  5.033s